requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
oauth2client==4.1.3
openpyxl==3.1.5
